ASoC: Intel: bytct-rt5640: add AIF_SWAP quirk
Some Baytrail devices are wired with a connection to AIF2 instead of
the more usual AIF1. Add ability to swap AIF1 and AIF2.

This quirk can only be used on regular hardware when using the
modified firmware which swaps SSP2 and SSP0. The audio functionality is
preserved by re-swapping the interfaces in the codec by setting the
controls "DAI select" and "SDI select" to 1.

The quirks are now handled in two phases since AIF name changes
need to be handled before card registration.