ASoC: Intel: Atom: auto-detection of Baytrail-CR
BYT-CR does not have SSP2 pins available and needs to use
SSP0 instead for codec connection. The upstream code needs to
be patched manually on BYT-CR platforms to load a dedicated
firmware that drives SSP0 internally.
For some reason the IPC interrupt index is also modified from
the Baytrail-T reference.

Use PUNIT BIOS config bits to infer which SST firmware needs to
needs to be loaded without user configuration. Assume regular
firmware is to be used if status is incorrect or CONFIG_IOSF_MBI
is not enabled.

Tested on Asus T100TA and T100TAF.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>